# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payproxy API",
        "description": "API pública para emissão de boletos bancários com split de pagamento. Autenticação via API Key no header `X-Api-Key`.",
        "contact": {
            "name": "Ciberian",
            "email": "suporte@ciberian.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Produção"
        }
    ],
    "paths": {
...
        "/webhooks/pjbank": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Webhook de pagamento PJBank",
                "description": "Endpoint chamado pelo PJBank para notificar pagamentos. Não é chamado pelos tenants. A autenticidade é validada via HMAC-SHA256 (header `X-PJBank-Hmac-Sha256`) e pela credencial do boleto. Retorna 200 mesmo em cenários de idempotência para evitar retentativas do PJBank.",
                "operationId": "697cc741e2e222b9c02f4e5107656ecd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nosso_numero": {
                                        "description": "Identificador do boleto no PJBank",
                                        "type": "string",
                                        "example": "99999999"
                                    },
                                    "token_transaction": {
                                        "description": "Token alternativo de identificação",
                                        "type": "string",
                                        "example": "tok_abc123"
                                    },
                                    "credencial": {
                                        "description": "Credencial do tenant no PJBank (validada fail-closed)",
                                        "type": "string",
                                        "example": "cred_xxxxx"
                                    },
                                    "valor": {
                                        "description": "Valor pago em reais",
                                        "type": "number",
                                        "example": 150
                                    },
                                    "forma_pagamento": {
                                        "description": "Canal de pagamento: barcode, pix, ted, etc.",
                                        "type": "string",
                                        "example": "pix"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notificação processada (ou ignorada por idempotência)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Assinatura HMAC inválida ou credencial incorreta/ausente"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "ApiKey": {
                "type": "apiKey",
                "description": "API Key gerada no backoffice Payproxy. Formato: `ppx_<40 chars>`.",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
...
}